refactor: reorganize contract and ABI structure for fitness agent
