+++ src/config/addresses.ts
@@ -1,0 +1,1 @@
+const FITNESS_AGENT_ADDRESS = "0x054ba199Ef61ef15226e2CeB61138f7d5E2F8408";
